rebase-interactive: show revert option and add single letter shortcut
A `revert` in an interactive rebase can be useful, e.g. if a faulty
commit was pushed to the main branch already, so you can't just drop it.
When you are already working in a feature branch you might just want to
revert said commit right where you branched off from main, so you can
continue working on the feature you intend while still being up-to-date
otherwise.

Another reason why you might not want to drop a commit is if it is a
work in progress one and you want to properly fix it later, but for now
need to revert the changes. That way it is a lot cleaner to structure
your branch like this:

    A---B---C       (B is WIP commit you cannot use as is)
=>
    A---B---~B---C  (temporarily revert B (called "~B") directly after
                     it is created, until you find the time to fix it -
                     at which point in time you will naturally drop the
                     revert commit)

This way you still have the WIP patch, but "your history is not broken
the whole time".

Signed-off-by: Michael Lohmann <mi.al.lohmann@gmail.com>